test pumpx_methods
